Refactor share booking out of booking service
