+++ src/main/java/uk/gov/hmcts/reform/preapi/services/BookingService.java
@@ -7,13 +7,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.preapi.dto.BookingDTO;
 import uk.gov.hmcts.reform.preapi.dto.CreateBookingDTO;
-import uk.gov.hmcts.reform.preapi.dto.ShareBookingDTO;
 import uk.gov.hmcts.reform.preapi.entities.Booking;
 import uk.gov.hmcts.reform.preapi.entities.Case;
 import uk.gov.hmcts.reform.preapi.entities.Participant;
-import uk.gov.hmcts.reform.preapi.entities.ShareBooking;
 import uk.gov.hmcts.reform.preapi.enums.UpsertResult;
-import uk.gov.hmcts.reform.preapi.exception.ConflictException;
 import uk.gov.hmcts.reform.preapi.exception.NotFoundException;
 import uk.gov.hmcts.reform.preapi.exception.ResourceInDeletedStateException;
 import uk.gov.hmcts.reform.preapi.repositories.BookingRepository;
@@ -37,8 +34,6 @@ public class BookingService {
 
     private final BookingRepository bookingRepository;
     private final ParticipantRepository participantRepository;
-    private final UserRepository userRepository;
-    private final ShareBookingRepository shareBookingRepository;
     private final CaseRepository caseRepository;
     private final CaptureSessionService captureSessionService;
 
@@ -46,13 +41,9 @@ public class BookingService {
     public BookingService(final BookingRepository bookingRepository,
                           final CaseRepository caseRepository,
                           final ParticipantRepository participantRepository,
-                          final UserRepository userRepository,
-                          final ShareBookingRepository shareBookingRepository,
                           CaptureSessionService captureSessionService) {
         this.bookingRepository = bookingRepository;
         this.participantRepository = participantRepository;
-        this.userRepository = userRepository;
-        this.shareBookingRepository = shareBookingRepository;
         this.caseRepository = caseRepository;
         this.captureSessionService = captureSessionService;
     }
@@ -151,29 +142,6 @@ public class BookingService {
         }
         captureSessionService.deleteCascade(entity.get());
         bookingRepository.deleteById(id);
-    }
-
-    @Transactional
-    public UpsertResult shareBookingById(ShareBookingDTO shareBookingDTO) {
-        if (shareBookingRepository.existsById(shareBookingDTO.getId())) {
-            throw new ConflictException("Share booking already exists");
-        }
-
-        final var booking = bookingRepository.findById(shareBookingDTO.getBookingId())
-            .orElseThrow(() -> new NotFoundException("Booking: " + shareBookingDTO.getBookingId()));
-        final var sharedByUser = userRepository.findById(shareBookingDTO.getSharedByUserId())
-            .orElseThrow(() -> new NotFoundException("Shared by User: " + shareBookingDTO.getSharedByUserId()));
-        final var sharedWithUser = userRepository.findById(shareBookingDTO.getSharedWithUserId())
-            .orElseThrow(() -> new NotFoundException("Shared with User: " + shareBookingDTO.getSharedWithUserId()));
-
-        var shareBookingEntity = new ShareBooking();
-        shareBookingEntity.setId(shareBookingDTO.getId());
-        shareBookingEntity.setBooking(booking);
-        shareBookingEntity.setSharedBy(sharedByUser);
-        shareBookingEntity.setSharedWith(sharedWithUser);
-        shareBookingRepository.save(shareBookingEntity);
-
-        return UpsertResult.CREATED;
     }
 
     @Transactional
