+++ src/test/java/uk/gov/hmcts/reform/preapi/services/BookingServiceTest.java
@@ -9,12 +9,10 @@ import org.springframework.data.domain.PageImpl;
 import uk.gov.hmcts.reform.preapi.dto.BookingDTO;
 import uk.gov.hmcts.reform.preapi.dto.CreateBookingDTO;
 import uk.gov.hmcts.reform.preapi.dto.CreateParticipantDTO;
-import uk.gov.hmcts.reform.preapi.dto.ShareBookingDTO;
 import uk.gov.hmcts.reform.preapi.entities.Booking;
 import uk.gov.hmcts.reform.preapi.entities.Case;
 import uk.gov.hmcts.reform.preapi.entities.Court;
 import uk.gov.hmcts.reform.preapi.entities.Participant;
-import uk.gov.hmcts.reform.preapi.entities.User;
 import uk.gov.hmcts.reform.preapi.enums.ParticipantType;
 import uk.gov.hmcts.reform.preapi.enums.UpsertResult;
 import uk.gov.hmcts.reform.preapi.exception.NotFoundException;
@@ -58,12 +56,6 @@ class BookingServiceTest {
 
     @MockBean
     private ParticipantRepository participantRepository;
-
-    @MockBean
-    private ShareBookingRepository shareBookingRepository;
-
-    @MockBean
-    private UserRepository userRepository;
 
     @MockBean
     private CaptureSessionService captureSessionService;
@@ -378,139 +370,6 @@ class BookingServiceTest {
 
         verify(captureSessionService, never()).deleteCascade(any(Booking.class));
         verify(bookingRepository, never()).deleteById(bookingId);
-    }
-
-    @DisplayName("Share a booking by its id")
-    @Test
-    void shareBookingSuccess() {
-        var shareBookingDTO = new ShareBookingDTO();
-        shareBookingDTO.setId(UUID.randomUUID());
-        shareBookingDTO.setBookingId(UUID.randomUUID());
-        shareBookingDTO.setSharedByUserId(UUID.randomUUID());
-        shareBookingDTO.setSharedWithUserId(UUID.randomUUID());
-
-        var bookingEntity = new Booking();
-        var sharedByUser = new User();
-        var sharedWithUser = new User();
-
-        when(
-            bookingRepository.findById(shareBookingDTO.getBookingId())
-        ).thenReturn(Optional.of(bookingEntity));
-        when(
-            userRepository.findById(shareBookingDTO.getSharedByUserId())
-        ).thenReturn(Optional.of(sharedByUser));
-        when(
-            userRepository.findById(shareBookingDTO.getSharedWithUserId())
-        ).thenReturn(Optional.of(sharedWithUser));
-
-        assertThat(bookingService.shareBookingById(shareBookingDTO)).isEqualTo(UpsertResult.CREATED);
-    }
-
-    @DisplayName("Share a booking by its id when booking doesn't exist")
-    @Test
-    void shareBookingFailureBookingDoesntExist() {
-        var shareBookingDTO = new ShareBookingDTO();
-        shareBookingDTO.setId(UUID.randomUUID());
-        shareBookingDTO.setBookingId(UUID.randomUUID());
-        shareBookingDTO.setSharedByUserId(UUID.randomUUID());
-        shareBookingDTO.setSharedWithUserId(UUID.randomUUID());
-
-        when(
-            bookingRepository.findById(shareBookingDTO.getBookingId())
-        ).thenReturn(Optional.empty());
-
-        assertThatExceptionOfType(uk.gov.hmcts.reform.preapi.exception.NotFoundException.class)
-            .isThrownBy(() -> {
-                bookingService.shareBookingById(shareBookingDTO);
-            })
-            .withMessage("Not found: Booking: " + shareBookingDTO.getBookingId());
-    }
-
-    @DisplayName("Share a booking by its id when shared by user doesn't exist")
-    @Test
-    void shareBookingFailureSharedByUserDoesntExist() {
-        var shareBookingDTO = new ShareBookingDTO();
-        shareBookingDTO.setId(UUID.randomUUID());
-        shareBookingDTO.setBookingId(UUID.randomUUID());
-        shareBookingDTO.setSharedByUserId(UUID.randomUUID());
-        shareBookingDTO.setSharedWithUserId(UUID.randomUUID());
-
-        var bookingEntity = new Booking();
-
-        when(
-            bookingRepository.findById(shareBookingDTO.getBookingId())
-        ).thenReturn(Optional.of(bookingEntity));
-        when(
-            userRepository.findById(shareBookingDTO.getSharedByUserId())
-        ).thenReturn(Optional.empty());
-
-        assertThatExceptionOfType(uk.gov.hmcts.reform.preapi.exception.NotFoundException.class)
-            .isThrownBy(() -> {
-                bookingService.shareBookingById(shareBookingDTO);
-            })
-            .withMessage("Not found: Shared by User: " + shareBookingDTO.getSharedByUserId());
-    }
-
-    @DisplayName("Share a booking by its id when shared with user doesn't exist")
-    @Test
-    void shareBookingFailureSharedWithUserDoesntExist() {
-        var shareBookingDTO = new ShareBookingDTO();
-        shareBookingDTO.setId(UUID.randomUUID());
-        shareBookingDTO.setBookingId(UUID.randomUUID());
-        shareBookingDTO.setSharedByUserId(UUID.randomUUID());
-        shareBookingDTO.setSharedWithUserId(UUID.randomUUID());
-
-        var bookingEntity = new Booking();
-        var sharedByUser = new User();
-
-        when(
-            bookingRepository.findById(shareBookingDTO.getBookingId())
-        ).thenReturn(Optional.of(bookingEntity));
-        when(
-            userRepository.findById(shareBookingDTO.getSharedByUserId())
-        ).thenReturn(Optional.of(sharedByUser));
-        when(
-            userRepository.findById(shareBookingDTO.getSharedWithUserId())
-        ).thenReturn(Optional.empty());
-
-        assertThatExceptionOfType(uk.gov.hmcts.reform.preapi.exception.NotFoundException.class)
-            .isThrownBy(() -> {
-                bookingService.shareBookingById(shareBookingDTO);
-            })
-            .withMessage("Not found: Shared with User: " + shareBookingDTO.getSharedWithUserId());
-    }
-
-    @DisplayName("Share a booking by its id when share booking already exists")
-    @Test
-    void shareBookingFailureShareBookingAlreadyExists() {
-        var shareBookingDTO = new ShareBookingDTO();
-        shareBookingDTO.setId(UUID.randomUUID());
-        shareBookingDTO.setBookingId(UUID.randomUUID());
-        shareBookingDTO.setSharedByUserId(UUID.randomUUID());
-        shareBookingDTO.setSharedWithUserId(UUID.randomUUID());
-
-        var bookingEntity = new Booking();
-        var sharedByUser = new User();
-        var sharedWithUser = new User();
-
-        when(
-            bookingRepository.findById(shareBookingDTO.getBookingId())
-        ).thenReturn(Optional.of(bookingEntity));
-        when(
-            userRepository.findById(shareBookingDTO.getSharedByUserId())
-        ).thenReturn(Optional.of(sharedByUser));
-        when(
-            userRepository.findById(shareBookingDTO.getSharedWithUserId())
-        ).thenReturn(Optional.of(sharedWithUser));
-        when(
-            shareBookingRepository.existsById(shareBookingDTO.getId())
-        ).thenReturn(true);
-
-        assertThatExceptionOfType(uk.gov.hmcts.reform.preapi.exception.ConflictException.class)
-            .isThrownBy(() -> {
-                bookingService.shareBookingById(shareBookingDTO);
-            })
-            .withMessage("Conflict: Share booking already exists");
     }
 
     @DisplayName("Should delete all attached capture sessions before marking booking as deleted")
